Incorrect Approach:
The code const tweetInput = tweetLink.elements.tweet.value attempts to access the value of the tweet input field directly. However, this is incorrect because:

Chaining Properties: You're trying to access a property (.value) of a non-existent property (.tweet).

tweetLink.elements is a collection of form elements, not a single element.
Missing Reference: You haven't established a reference to the specific tweet input field within the form.